requirements.txt
pillow == 11.0.0